Separate utils to more precise files
